graph.py: Change isomorphism to monomorphism
Due to a different definition in NetworkX, what we are seeking isn't
a isomorphism, but a monomorphism. When NetworkX says 'subgraph of A
is isomorphic to B' it means that the node-induced subgraph of A is
*exactly* the same as B. What it really means that is that if you
add one more edge between these nodes of A which are said to be
isomorphic, then nx doesn't find it, as the node-induced subgraph isn't
*exactly* the same. For instance, if we can't apply production P16
to a star with 5 leaves and edge between vert 1 and 2 if we choose to
use isomorphisms.

On the contrary, when NetworkX says 'subgraph of A is monomorphic to B'
what it really means is that the found node-induced subgraph of A
*contains* graph B in it. In other words, it allows additional edges
beyond those that are in graph B.

Further explaination about definition used in NetworkX can be found at:
https://networkx.org/documentation/stable/reference/algorithms/isomorphism.vf2.html#subgraph-isomorphism

Unfortunately, there is no documentation for monomorphisms, but it
is implemented in the source code:
https://networkx.org/documentation/stable/_modules/networkx/algorithms/isomorphism/isomorphvf2.html