model.py: Add support for pentagons
Introduce new node label `p', which indicates that the node is the
hyperedge for pentagon. In that case attribute `hanging' has the same
meaning as with label `q', which means that the hyperedge has been
chosen to be broken down.

util.py: Add primitives for creating graphs

Add support for creating:
 * square;
 * star with 5 leaves;
 * pentagon.

Square is created with `q' node inside, while star and pentagon
with `p' node. Parameter `for_lhs' changes the enumeration of
nodes; if set, the numeration start from 0.