graph.py: Fix the edge_equality (#27)
Matching edges exactly with flag attributes causes problems with border.
Suppose we have split the square grid many times and one square lies
partly on the border. Then the productions which we have won't catch it
because flag value differs from lhs. So in order to catch these situations we
would have to explicitly introduce 4 new productions (1 border, 2 borders
etc) in case of squares and in case of pentagons it's even more.

Since all relevant information is stored in vertices, after matching
nodes exactly we don't have to match edges exactly; that's why we fix it
by removing second condition.

Signed-off-by: Michał Grzelak <mchl.grzlk@gmail.com>